Move some generic funtions to util.py
